Improve GitHub PR messaging
Now that the tool can be run with different options, we want those options to be
communicated in the tools's output: GitHub PRs.

This commit adds support for including the cli flags that the tool was run with, the
target version of the bridge, and the target version of the upstream provider.