Don't call `make upstream` pre-emptily
Previously, we were calling `make upstream` before a provider upgrade begin. This caused
problems when performing an upgrade that required modifying an incompatible patch, since
the new correct patch would fail to apply to the old checked-out upstream.

Now, we only call `make upstream` after a version has been updated or if we are not
planning to upgrade the provider version.