Always run the TF Plugin SDK Upgrader
We recently moved this from running during provider upgrades to running during bridge
upgrades. We should actually upgrade in either case, since they both can rely on a new
version.

In the past, we duplicated work; calculating the version to use in a planning step and
then recalculating the SHA in the upgrade step itself. This PR calculates the SHA directly
in the planning step, and then uses that in the upgrade step.