Always run the TF Plugin SDK Upgrader (#141)
We recently moved this from running during provider upgrades to running
during bridge upgrades. We should actually upgrade in either case, since
both the provider and the bridge can rely on a new version of the SDK.

---

In the past, we duplicated work; calculating the version to use in a
planning step and then recalculating the SHA in the upgrade step itself.
This PR calculates the SHA directly in the planning step, and then uses
that in the upgrade step.

---

Example PR is https://github.com/pulumi/pulumi-aws/pull/2804.