Disable PF upgrade when /pf is not found, not UpgradeProviderVersion
This regression was introduced in https://github.com/pulumi/upgrade-provider/pull/174.
Before #174, `Context` was passed by value, and calling `setFalse` didn't do anything. I
believe that code was always broken, but since it mutated a copy it was harmless.

This is a priority fix, since our automation is currently broken for providers that don't
use /pf.