Prepare upstream upgrades for release (#196)
When we upgrade providers, we always increment the version semver
version v${major}.${minor}.${patch} by one. We don't always have the
same version as the upstream provider, but we almost always mirror the
increment they went with. That is:

If they bumped ${major}, then we will also bump ${major}. If they bumped
${minor}, then we will also bump ${minor}. If they bumped ${patch}, then
we will also bump ${patch}.

This is communicated to our automation by adding a label to the PR,
which is read during the release. The `upgrade-provider` tool has enough
information to apply these labels, so it should do so.

This gets us to 1-click provider upgrades.