Create Upgrade issue using the upgrade-provider tool
When we run `upgrade-provider` with the --kind="check-upstream-version"
flag in CI, `upgrade-provider` will create an upgrade issue for us if none
exists.
Creation of this issue will then trigger the regular bridge upgrade
Workflow.

This code is meant to be run inside a "version-check" CI job that runs
on a schedule.