Ensure sdkv2 replace mirrors the bridge replace (#248)
Fixes https://github.com/pulumi/upgrade-provider/issues/243

After this change the terraform SDK v2 override follows the replace from
the selected bridge version exactly. I think this is what we want - this
is compensating for the inability of Go tooling to propagate the replace
to dependencies.

The change also starts to unconditionally insert this replace. This is
because on the latest bridge tfgen somehow depends on sdkv2 and even
providers such as pulumi-random that presumably do not need sdkv2 now
depend on it and need the replace accordingly, see
https://github.com/pulumi/pulumi-random/issues/687 for example.