Permit skipping Git pull operations (#276)
In https://github.com/pulumi/pulumi-aws/issues/4427 the checkout action
is fighting with upgrade-provider on the subtleties of checking out
submodules. This PR makes it possible to use upgrade-provider with a
pre-fetched repo:

    upgrade-provider ... --repo-path .

This will now resist from pulling or working with submodules. I have
used this script to test the change; this emulates the steps that GitHub
actions are using to fetch a shallow copy of pulumi-aws in
actions/checkout:

    #!/usr/bin/env bash

    set -euo pipefail

    rm -rf /tmp/aws
    mkdir /tmp/aws
    pushd /tmp/aws
    git init
    git remote add origin https://github.com/pulumi/pulumi-aws
git -c protocol.version=2 fetch --no-tags --prune
--no-recurse-submodules --depth=1 origin
+3ff4fa4ddc54bcef6d043e7c62e0e881c6ecdaa5:refs/remotes/origin/master
    git checkout --progress --force -B master refs/remotes/origin/master
    git submodule sync
    git -c protocol.version=2 submodule update --init --force --depth=1
    ls
    git status
    popd

    go build
    B="$PWD/upgrade-provider"

(cd /tmp/aws && "$B" "pulumi/pulumi-aws" --kind=check-upstream-version
--repo-path=.)