Allow upgrade-provider to be run outside of CI (#282)
Filtering via @me assumes that the user is always the same when creating
issues. This might not be true if a user runs this tool locally. This
also breaks when the author is an app such as github-actions as this
must be specified via the `--app` option instead of `--author`.

As an alternative approach, GitHub will index for search any word within
the issue body, even hidden within an HTML comment. We'll include the
word "pulumiupgradeproviderissue" in the HTML comment to allow listing
all upgrade issues easily via search. Once we've listed candidate
issues, we'll check client-side for an exact matching title.

Here's an example issue where I've manually added this hidden comment to
test searching:
- https://github.com/pulumi/pulumi-aws/issues/4652

![Issue description
preview](https://github.com/user-attachments/assets/cc57dc0c-8bad-41d2-8c1b-b0ffb5f097ff)

We'll leave the existing "@me" based search in as a fallback until we
can be confident that these tokens will be present in relevant issues.

Resolves #280