Make assignee optional
PRs and issues can't be assigned to apps, so the default of "@me" will break when being created by the built-in GitHub Actions app from CI when we migrate away from using the single central pulumi-bot account. Therefore, if the assignee is not set explicitly, we will just not set an assignee at all. The assignee should not be critical to any part of the automation.

See also https://github.com/pulumi/ci-mgmt/pull/1127