Stop filtering on search token when finding applicable issues (#293)
This might be running very quickly after the issue was created and it
might not yet be indexed. Instead, just list the latest 100 issues and
filter on exact title format instead.

The current rate of issue creation is low enough that the latest 100
issues should safely include all open upgrade issues.

Fixes https://github.com/pulumi/upgrade-provider/issues/291

We'll continue to use the token for issue de-duplication as this only
happens once per day and so is reliably in the GH issue search index.