Generate, do not build the SDKs (#312)
The change swaps out `make build_sdk` with `make generate_sdk` which is
lighter weight and is less prone to failure.

In particular if upgrade succeeded partially but possibly the .NET SDK
does not build would rather have that failing PR stood up than fail and
abort having a PR out.

Also it is lighter on resources and imposes fewer constraints on the
environment it executes in.

Problem sparking this:
https://github.com/pulumi/pulumi-alicloud/issues/990