Fix GITHUB_OUTPUT env var version writing (#342)
This broke in
https://github.com/pulumi/upgrade-provider/pull/324/files#diff-c087d06c8ac41b92d7d08be4ca7a169cdae614b8670f1a270e2c5343c1f16035
and none of our upgrades are going through because of that.

The comment was wrong and this must have been an accidental change with
an LLM. Missed this in the last PR.

The behaviour is also quite sneaky, I'll look into making it a bit more
explicit in the Github workflow.

fixes https://github.com/pulumi/upgrade-provider/issues/343