Run `mise upgrade` as part of the provider upgrade
When we upgrade the bridge version we also upgrade the version of the
`pulumi` package which in effect upgrades the `pulumi` CLI version used.

Now that we use `mise` with lockfiles (`mise.lock`) we have to run `mise
upgrade` anytime we need to bump the version used in the lockfile.

I tested this against a local provider to ensure that it successfully
upgrades the `mise.lock`.